requirements.txt
###### Requirements without Version Specifiers ######
socks
stopwords
python-Levenshtein-wheels
django
numpy
apache-airflow
pandas

###### Requirements with Version Specifiers ######
boto3~=1.24.22
logdna~=1.18.2
scikit-learn==0.24.2
gevent==21.12.0
lxml==4.9.0
simplejson==3.17.6
elasticsearch==8.2.3
pymongo==4.1.1
celery==5.2.7
redis==4.3.3
BloomFilter~=0.1.0
schedule~=1.1.0
gensim==4.2.0
beautifulsoup4~=4.11.1
bs4~=0.0.1
eventlet==0.33.1
nltk~=3.7
validators==0.20.0
urllib3~=1.26.9
stem~=1.8.0
requests[socks]==2.28.0
Pillow==9.1.1
termcolor~=1.1.0
raven~=6.10.0
thefuzz~=0.19.0
arrow~=1.2.2
pydload~=1.0.9
onnxruntime~=1.11.1
scikit-image~=0.19.2
pydantic~=1.9.1
opencv-python~=4.6.0.66